[mathjax-3] Change Mathjax runtime to 3.0 (#315)
