+++ react-renderer/package.json
@@ -9,6 +9,7 @@
     "debug": "^4.1.1",
     "fast-memoize": "^2.5.1",
     "lodash": "^4.17.13",
+    "mathjax-full": "^3.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-fast-compare": "^2.0.4",
