+++ react-renderer/public/index.html
@@ -13,10 +13,7 @@
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-    <script
-      type="text/javascript"
-      src="%PUBLIC_URL%/MathJax.min.js?delayStartupUntil=configured"
-    ></script>
+
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
