+++ react-renderer/src/Util.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 
-declare const MathJax: any;
 import memoize from "fast-memoize";
 
 export const arrowheads = {
@@ -187,93 +186,6 @@ export const getAngle = (x1: number, y1: number, x2: number, y2: number) => {
 
 export const getLen = (x1: number, y1: number, x2: number, y2: number) => {
   return Math.sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
-};
-
-export function svgBBox(svgEl: SVGSVGElement) {
-  const tempDiv = document.createElement("div");
-  tempDiv.setAttribute(
-    "style",
-    "position:absolute; visibility:hidden; width:0; height:0"
-  );
-  document.body.appendChild(tempDiv);
-  const tempSvg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-  const tempG = document.createElementNS("http://www.w3.org/2000/svg", "g");
-  tempSvg.appendChild(tempG);
-  tempDiv.appendChild(tempSvg);
-  const tempEl = svgEl.cloneNode(true) as SVGSVGElement;
-  tempG.appendChild(tempEl);
-  const bb = tempG.getBBox();
-  document.body.removeChild(tempDiv);
-  return bb;
-}
-
-const tex2svg = memoize(
-  async (contents: string, name: string, fontSize: string): Promise<any> =>
-    new Promise(resolve => {
-      const wrapper = document.createElement("div");
-      // HACK: Style compiler decides to give empty labels if not specified
-      if (contents !== "") {
-        wrapper.innerHTML = "$" + contents + "$";
-        MathJax.Hub.Queue(["Typeset", MathJax.Hub, wrapper]);
-        MathJax.Hub.Queue(() => {
-          const output = wrapper.getElementsByTagName("svg")[0];
-          output.setAttribute("xmlns", "http://www.w3.org/2000/svg");
-          output.setAttribute("style", `font-size: ${fontSize}`);
-          // TODO: need to check whether MathJax returns a non-null response
-          // NOTE: This is where you can directly control the width/height of the LaTeX
-          const { width, height } = svgBBox(output);
-
-          // TODO: These seem to be extraneous. Remove?
-          // output.setAttribute("width", width.toString());
-          // output.setAttribute("height", height.toString());
-          const body = output;
-          resolve({ body, width, height });
-        });
-      } else {
-        resolve({ output: undefined, width: 0, height: 0 });
-      }
-      wrapper.remove();
-    })
-);
-
-// https://stackoverflow.com/a/44564236
-export const collectLabels = async (allShapes: any[]) => {
-  MathJax.Hub.Config({
-    skipStartupTypeset: true,
-    extensions: ["tex2jax.js", "TeX/AMSmath.js"],
-    jax: ["input/TeX", "output/SVG"],
-    // https://docs.mathjax.org/en/v2.7-latest/options/output-processors/SVG.html
-    SVG: {
-      matchFontHeight: false,
-      useGlobalCache: false, // Needed for SVG inline export
-      useFontCache: false // further reduces the reuse of paths
-    },
-    tex2jax: {
-      inlineMath: [["$", "$"], ["\\(", "\\)"]],
-      processEscapes: true
-    }
-  });
-  return Promise.all(
-    allShapes.map(async ([type, obj]: [string, any]) => {
-      if (type === "Text" || type === "TextTransform") {
-        const { body, width, height } = await tex2svg(
-          obj.string.contents,
-          obj.name.contents,
-          obj.fontSize.contents
-        );
-        // Instead of directly overwriting the properties, cache them temporarily and let `propogateUpdate` decide what to do
-        const obj2 = { ...obj };
-        obj2.w.updated = width;
-        obj2.h.updated = height;
-        // console.log(body, obj2);
-        // Add omit: true flag so it doesn't get sent to the server
-        obj2.rendered = { contents: body, omit: true };
-        return [type, obj2];
-      } else {
-        return [type, obj];
-      }
-    })
-  );
 };
 
 export const loadImageElement = memoize(
